[WIP] Split out codegen
Summary:

Test Plan:
