+++ src/utils/helpers.ts
@@ -50,6 +50,12 @@ export function nullThrows<T>(value: T | null | undefined): T {
   return value;
 }
 
+// Predicate function for filtering out null values
+// Includes TypeScript refinement for narrowing the type
+export function isNonNull<T>(value: T | null | undefined): value is T {
+  return value != null;
+}
+
 // Noop tagged template literal
 export function graphql(strings: TemplateStringsArray, ...values: any[]) {
   // Combine the template literal parts into a single string
